# hash_info: use attrs slotted class