--- a/setup.cfg
+++ b/setup.cfg
@@ -30,6 +30,7 @@ install_requires=
     dvc-objects==0.1.11
     diskcache>=5.2.1
     nanotime>=0.5.2
+    attrs>=21.3.0
 
 [options.extras_require]
 cli =