@skshetry 2022-08-14
I also tested with `mypy` on the previous version of our slotted class and the initialization is twice as fast:
```console
------------------------------------------------------------------------------------ benchmark 'test_hashinfo_create[HashInfo]': 3 tests ------------------------------------------------------------------------------------
Name (time in ns)                                    Min                   Max                Mean             StdDev              Median               IQR              Outliers  OPS (Mops/s)            Rounds  Iterations
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_hashinfo_create[HashInfo] (0001_mypyc)      73.0051 (1.0)      5,195.4799 (3.16)      80.4915 (1.0)      63.9627 (1.26)      75.4801 (1.0)      0.4348 (1.0)      1141;10672       12.4237 (1.0)       57147         200
test_hashinfo_create[HashInfo] (0002_attrs)     140.6468 (1.93)     2,156.5296 (1.31)     154.4964 (1.92)     50.8068 (1.0)      147.0300 (1.95)     4.2654 (9.81)     7964;16120        6.4726 (0.52)     198216          34
test_hashinfo_create[HashInfo] (0003_slots)     148.5759 (2.04)     1,643.7270 (1.0)      158.3347 (1.97)     53.8006 (1.06)     152.2421 (2.02)     1.7568 (4.04)     2783;16239        6.3157 (0.51)     197981          33
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

I did not see any changes with attrs + mypyc, maybe it's not supported or is too dynamic for mypyc to handle.